requirements.txt
Django==5.0.2
tensorflow==2.16.1
numpy==1.24.3
Pillow==10.0.0
requests==2.31.0
python-dotenv==1.0.1 